+++ public/api/post.js
@@ -2,7 +2,7 @@
 const BASE_URL = 'http://localhost:8080/v1/posts';
 
 /// 인증 요청
-import { authFetch } from './base.js';
+import { authFetch } from './authFetch.js';
 
 /// 등록
 export async function createPost(postData) {
